Fix for when the mbstring extension is not availabe.
Although it is hard to know without any tests or detailed history,
I think all of our mb_internal_encoding calls are defensive calls
for the situation where the user of the library has set some other
encoding. As such, disabling them when the extension is not
available should be safe.